Fix Issue https://github.com/socketstream/socketstream/issues/609
Call stream(plan.httpServer); after running tasks, such that responders are defined when the WebSocket layer is loaded. Pass the plan as an argument to tasks.defaults(plan); such that the decision to include the start-server task in the load-socketstream task can be based on plan.httpServer instead of ss.server.httpServer.